backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
chromadb>=0.4.22
openai>=1.30.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
python-docx>=1.1.0
PyPDF2>=3.0.0
mammoth>=1.8.0
